common: syna: resolve build warnings
Reviewed-on: https://gerrit-sha.synaptics.com/gerrit/280652

+++ arch/arm/mach-synaptics/drivers/video/load_logo.c
@@ -23,6 +23,8 @@
  */
 
 #include <linux/types.h>
+#include <cpu_func.h>
+#include <asm/cache.h>
 #include <stdlib.h>
 #include "vpp_api.h"
 #include "vbuf.h"
@@ -275,7 +277,7 @@ int syna_load_logo_info(int width, int height, VBUF_INFO *p_vpp_buf, FASTLOGO_IN
 	logo_buffer = fl_header->offset + header;
 	logo_size = (fl_header->stride * fl_header->height) + pad_size;
 	read_buffer = (UINT8 *)malloc(logo_size);
-	flush_dcache_range(img_buff, img_buff + read_size);
+	flush_dcache_range((unsigned long)img_buff, (unsigned long)(img_buff + read_size));
 	memcpy(read_buffer, logo_buffer, logo_size);
 
 	p_vpp_buf->m_srcfmt = LOGO_SRC_FMT;
